2022_0701_1134am_52

+++ stusb_4500_in_c/st45_3b_convert_config_to_nvm_buf.c
@@ -2,12 +2,19 @@
 
 
 ST45config* _st45_convert_config_to_nvm_buf(ST45config* ___st45config) {
-    uint8_t __u01 ;
+    uint16_t __u01 ;
+    //uint16_t __u02 ;
     if ( NULL == ___st45config ) return NULL ;
 
     ___st45config -> buf[3][2] &= (~0b110) ;
     __u01 = 0b110 & ((___st45config -> pdoAmount)<<1) ;
     ___st45config -> buf[3][2] |= __u01;
+
+    //PDO1v - fixed at 5V and is unable to change, ignore
+    //PDO2v : Load voltage (10-bit)
+    // -bit 9:2 - sector 4, byte 1, bits 0:7
+    // -bit 0:1 - sector 4, byte 0, bits 6:7
+
 
     return ___st45config ;
 } // _st45_convert_config_to_nvm_buf
